refactored env_bool & env_true from prettier into utils; movedd tests to test_utils.py
